Safari-611.1.10/LayoutTests/fast/selectors/lang-extended-filtering-with-string-arguments.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
    <div lang="en"></div>
    <div lang=""></div>
    <div lang="\"\""></div>
    <div lang="de-1997-Latn"></div>
    <script>
    description('Verify selector specifying extended filtering of :lang() pseudo class with string arguments');
    shouldBe('document.querySelectorAll(\':lang("en")\').length', '1');
    shouldBe('document.querySelectorAll(\':lang("de-1997")\').length', '1');
    shouldBe('document.querySelectorAll(\':lang("*-1997")\').length', '1');
    shouldBe('document.querySelectorAll(":lang(\'\')").length', "0");
    shouldBe('document.querySelectorAll(\':lang("")\').length', '0');
    shouldBe('document.querySelectorAll(\':lang("en", "*-1997")\').length', '2');
    shouldBe('document.querySelectorAll(\':lang("", "*-1997")\').length', '1');

    shouldThrow('document.querySelectorAll(\':lang("""")\')');
    </script>
    <script src="../../resources/js-test-post.js"></script>
</body>
</html>
